requirements.txt
Django<3.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
pytest-freezegun==0.4.2
